Rename cmdr to commander for approachability
