+++ tools/bin/modtool
@@ -7,8 +7,8 @@ import intelhex
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from pymomo.cmdr.meta import *
-from pymomo.cmdr.exceptions import *
+from pymomo.commander.meta import *
+from pymomo.commander.exceptions import *
 import cmdln
 from colorama import Fore, Style
